Fix broken handling of invalid binaries on OS X

+++ mozmill_automation/testrun.py
@@ -53,6 +53,7 @@ class TestRun(object):
             self.repository_url = self.options.repository_url
         else:
             self.repository_url = MOZMILL_TESTS_REPOSITORIES[self.options.application]
+        self._repository = None
 
         self.addon_list = []
         self.downloaded_addons = []
@@ -181,7 +182,10 @@ class TestRun(object):
                     if sys.platform == "darwin":
                         # Ensure that self._folder is the app bundle on OS X
                         p = re.compile('.*\.app/')
-                        self._folder = p.search(self.binary).group()
+                        try:
+                            self._folder = p.search(self.binary).group()
+                        except:
+                            raise errors.InvalidBinaryException(self.binary)
                     else:
                         self._folder = os.path.dirname(self.binary)
 
@@ -268,7 +272,8 @@ class TestRun(object):
             self.remove_downloaded_addons()
 
             # Remove the temporarily cloned repository
-            self._repository.remove()
+            if self._repository:
+                self._repository.remove()
 
             # If a test has been failed ensure that we exit with status 2
             if self.last_failed_tests:
